+++ Modules/HR/Http/Controllers/Recruitment/ApplicationController.php
@@ -327,4 +327,5 @@ abstract class ApplicationController extends Controller
 
         return redirect(route('applications.job.index'))->with('status', $status);
     }
+
 }
