+++ Modules/HR/Services/ApplicationService.php
@@ -102,4 +102,49 @@ class ApplicationService implements ApplicationServiceContract
         $data->actual_end_time = $meetDuration;
         $data->save();
     }
+
+    public function getVolunteerChartData()
+    {
+        $jobs = Job::where('type', 'volunteer')->get('id');
+
+        $data = [];
+        foreach ($jobs as $job) {
+            $data[] = $job->id;
+        }
+        $todayCount = Application::whereDate('created_at', now())->whereIn('hr_job_id', $data)
+            ->count();
+        $record = Application::select(
+            \DB::raw('COUNT(*) as count'),
+            \DB::raw('MONTHNAME(created_at) as month_created_at'),
+            \DB::raw('DATE(created_at) as date_created_at')
+        )
+            ->where('created_at', '>', Carbon::now()->subDays(23))
+            ->whereIn('hr_job_id', $data)
+            ->groupBy('date_created_at', 'month_created_at')
+            ->orderBy('date_created_at', 'ASC')
+            ->get();
+
+        $data = [];
+        foreach ($record as $row) {
+            $data['label'][] = (new Carbon($row->date_created_at))->format('M d');
+            $data['data'][] = (int) $row->count;
+            $verifiedApplications = Application::where('is_verified', 1)->whereDate('created_at', '=', date(new Carbon($row->date_created_at)))->count();
+            $data['afterBody'][] = $verifiedApplications;
+        }
+        $data['chartData'] = json_encode($data);
+
+        return $data['chartData'];
+    }
+
+    public function TodayVolunteerCount()
+    {
+        $data = [];
+        $jobs = Job::where('type', 'volunteer')->get('id');
+        foreach ($jobs as $job) {
+            $data[] = $job->id;
+        }
+        $todayCount = Application::whereDate('created_at', now())->whereIn('hr_job_id', $data)
+            ->count();
+        return $todayCount;
+    }
 }
