+++ routes/web.php
@@ -11,7 +11,7 @@
 |
  */
 
-use Modules\HR\Http\Controllers\Recruitment\ApplicantController;
+use App\Http\Controllers\HR\Applications\VolunteerApplicationController;
 
 Route::get('/', function () {
     if (Auth::check()) {
@@ -167,5 +167,5 @@ Route::middleware('auth')->group(function () {
     Route::get('user/read-books', 'UserBookController@index');
     Route::get('user/wishlist-books', 'UserBookController@booksInWishlist');
     Route::get('user/projects', 'UserController@projects');
-    Route::get('DailyCountVolunteers', [ApplicantController::class, 'findVolunteerApplicant'])->name('volunteers.dailyapplicationscount');
+    Route::get('DailyCountVolunteers', [VolunteerApplicationController::class, 'findVolunteerApplicant'])->name('volunteers.dailyapplicationscount');
 });
